requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
